backend/requirements.txt
Django>=3.2.0,<3.3.0
django-cors-headers
django-filter>=2.4.0,<3.0.0
djangorestframework>=3.12.0,<3.13.0
Pillow
psycopg2-binary
cryptography==3.4.8
gunicorn==20.1.0
gunicorn[gevent]
djoser
djangorestframework_simplejwt
sentry-sdk
django-simple-history
kavenegar
python-memcached
ffmpeg-python
django-import-export
